refactor: update import paths for AudioPlayButton component
